application/requirements.txt
fastapi
uvicorn
requests
pydantic
motor
datetime
typing
bcrypt
cryptography
passlib
jose

python-dotenv